Parse messages from items healing a Pokemon
An addition to e93835abb1419bc28fb9e2f342ab563f5428ac04. This allows the
environment to parse items that heal the holder: Most notably Leftovers
and berries.

This also required an UNKNOWN_ITEM constant that each Pokemon's item
will now be initialized as, allowing the environment to distinguish between no item
(None) and an unknown item. This is necessary because the PS server will
give a log that looks like this when a berry is consumed:

```
|-damage|p2a: Blissey|22/714
|-enditem|p2a: Blissey|Sitrus Berry|[eat]
|-heal|p2a: Blissey|200/714|[from] item: Sitrus Berry
```

In the above log, the berry has already been consumed, but the heal
statement comes afterwards. The item will be None in this situation
because the `-enditem` log will have set it to None.